src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/shadow-root-text-child-expected.txt

CONSOLE ERROR: line 10: Uncaught TypeError: root.createShadowRoot is not a function
